refactor: move restore player state to standard
