+++ Mine3Auth/pom.xml
@@ -81,10 +81,6 @@
             <id>sonatype</id>
             <url>https://oss.sonatype.org/content/groups/public/</url>
         </repository>
-        <repository>
-            <id>gravemc-repo</id>
-            <url>https://repo.gravemc.net/releases/</url>
-        </repository>
     </repositories>
 
     <dependencies>
@@ -99,12 +95,6 @@
             <artifactId>jedis</artifactId>
             <version>5.1.2</version>
         </dependency>
-        <dependency>
-            <groupId>dev.iiahmed</groupId>
-            <artifactId>ModernDisguise</artifactId>
-            <version>3.1</version>
-            <scope>compile</scope>
-        </dependency>
         <!-- https://mvnrepository.com/artifact/io.javalin/javalin -->
         <dependency>
             <groupId>io.javalin</groupId>
@@ -116,12 +106,6 @@
             <groupId>org.web3j</groupId>
             <artifactId>crypto</artifactId>
             <version>4.12.0</version>
-        </dependency>
-        <!-- https://mvnrepository.com/artifact/org.postgresql/postgresql -->
-        <dependency>
-            <groupId>org.postgresql</groupId>
-            <artifactId>postgresql</artifactId>
-            <version>42.7.3</version>
         </dependency>
         <dependency>
             <groupId>org.slf4j</groupId>
