+++ Mine3Auth/src/main/java/com/boomchanotai/mine3Auth/config/Config.java
@@ -8,11 +8,6 @@ public class Config {
     public static String REDIS_HOST;
     public static int REDIS_PORT;
 
-    // Postgres Settings
-    public static String POSTGRES_HOST;
-    public static String POSTGRES_USERNAME;
-    public static String POSTGRES_PASSWORD;
-
     // Httpserver Settings
     public static int HTTPSERVER_PORT;
 
@@ -49,10 +44,6 @@ public class Config {
 
         REDIS_HOST = config.getString("redis.host");
         REDIS_PORT = config.getInt("redis.port");
-
-        POSTGRES_HOST = config.getString("postgres.host");
-        POSTGRES_USERNAME = config.getString("postgres.username");
-        POSTGRES_PASSWORD = config.getString("postgres.password");
 
         HTTPSERVER_PORT = config.getInt("httpserver.port");
