+++ Mine3Standard/src/main/java/com/boomchanotai/mine3Standard/commands/BanCommand.java
@@ -1,4 +1,4 @@
-package com.boomchanotai.mine3Auth.commands;
+package com.boomchanotai.mine3Standard.commands;
 
 import java.util.Date;
 
@@ -7,7 +7,7 @@ import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
-import com.boomchanotai.mine3Auth.logger.Logger;
+import com.boomchanotai.mine3Standard.logger.Logger;
 import com.boomchanotai.mine3Lib.address.Address;
 import com.boomchanotai.mine3Lib.repositories.PlayerRepository;
