+++ Mine3Standard/src/main/java/com/boomchanotai/mine3Standard/commands/PardonCommand.java
@@ -1,12 +1,12 @@
-package com.boomchanotai.mine3Auth.commands;
+package com.boomchanotai.mine3Standard.commands;
 
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
-import com.boomchanotai.mine3Auth.logger.Logger;
-import com.boomchanotai.mine3Auth.repositories.PostgresRepository;
+import com.boomchanotai.mine3Standard.logger.Logger;
+import com.boomchanotai.mine3Standard.repositories.PostgresRepository;
 import com.boomchanotai.mine3Lib.address.Address;
 import com.boomchanotai.mine3Lib.repositories.PlayerRepository;
