+++ Mine3Standard/src/main/java/com/boomchanotai/mine3Standard/entities/PlayerLocation.java
@@ -1,4 +1,4 @@
-package com.boomchanotai.mine3Auth.entities;
+package com.boomchanotai.mine3Standard.entities;
 
 import org.bukkit.World;
