+++ Mine3Standard/src/main/java/com/boomchanotai/mine3Standard/repositories/ItemStackAdapter.java
@@ -1,4 +1,4 @@
-package com.boomchanotai.mine3Auth.repositories;
+package com.boomchanotai.mine3Standard.repositories;
 
 import java.sql.SQLException;
 import java.util.ArrayList;
@@ -9,7 +9,7 @@ import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.ItemMeta;
 import org.postgresql.util.PGobject;
 
-import com.boomchanotai.mine3Auth.logger.Logger;
+import com.boomchanotai.mine3Standard.logger.Logger;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.JsonNode;
