+++ Mine3Standard/src/main/java/com/boomchanotai/mine3Standard/repositories/PotionEffectAdapter.java
@@ -1,4 +1,4 @@
-package com.boomchanotai.mine3Auth.repositories;
+package com.boomchanotai.mine3Standard.repositories;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -8,7 +8,7 @@ import org.bukkit.configuration.serialization.ConfigurationSerialization;
 import org.bukkit.potion.PotionEffect;
 import org.postgresql.util.PGobject;
 
-import com.boomchanotai.mine3Auth.logger.Logger;
+import com.boomchanotai.mine3Standard.logger.Logger;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
