+++ Mine3Standard/src/main/java/com/boomchanotai/mine3Standard/tabcompletion/PardonTabCompletion.java
@@ -1,4 +1,4 @@
-package com.boomchanotai.mine3Auth.commands;
+package com.boomchanotai.mine3Standard.tabcompletion;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -7,7 +7,7 @@ import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 import org.bukkit.command.TabCompleter;
 
-import com.boomchanotai.mine3Auth.repositories.PostgresRepository;
+import com.boomchanotai.mine3Standard.repositories.PostgresRepository;
 import com.boomchanotai.mine3Lib.address.Address;
 
 public class PardonTabCompletion implements TabCompleter {
