+++ Mine3Standard/src/main/resources/config.yml
@@ -3,6 +3,13 @@
 # Repository: https://github.com/boomchanotai/Mine3
 # Author: Boom Chanotai (https://www.boomchanotai.com/)
 
+# Postgres Settings
+postgres:
+  host: "jdbc:postgresql://localhost:5432/postgres"
+  username: "postgres"
+  password: "password"
+
+# Plugin Settings
 # {displayname} will be replaced with the player's display name with prefix and suffix group
 # {message} will be replaced with the message
 chat_format: "{displayname} &f: {message}"
