+++ Mine3Standard/tempCodeRunnerFile.xml
@@ -1,0 +1,7 @@
+
+        <dependency>
+            <groupId>dev.iiahmed</groupId>
+            <artifactId>ModernDisguise</artifactId>
+            <version>3.1</version>
+            <scope>compile</scope>
+        </dependency>
